requirements.txt
Flask==2.0.2
pytesseract==0.3.8
torch==1.10.0
opencv-python==4.5.4.60
numpy==1.26.4
Pillow==8.4.0
ultralytics=8.2.75
pandas=2.2.2
matplotlib==3.5.0
gitpython>=3.1.30
Pillow>=9.4.0
setuptools>=65.5.1
wheel>=0.38.0
scipy==1.7.3
scikit-learn==1.0.1
transformers=4.10.0
imutils==0.5.4
pyttsx3==2.90
